justi-fi-backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.10
langgraph>=0.0.10
groq>=0.4.0
chromadb>=0.4.0
python-dotenv>=0.19.0
requests>=2.26.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.2
torch>=2.0.0
notion-client>=2.2.0
python-dateutil>=2.8.2 